double instead of int

+++ lib/data/water_container.dart
@@ -2,7 +2,7 @@
 
 class WaterContainer {
   final int id;
-  final int currentWaterLevel;
+  final double currentWaterLevel;
 
   const WaterContainer({
     required this.id,
